+++ common/cfm2.h
@@ -1,9 +1,10 @@
-#ifndef CFM2_H
-#define CFM2_H
+#ifndef CFM_H
+#define CFM_H
     #include <stdio.h>
     #include <windows.h>
-    #include <direct.h>
-    #include <stringManipulation.h>
+    #include <direct.h>    
+    #include <sys/stat.h>
+    #include "stringManipulation.h"
 
     FILE* lookInRoot(char* filename, char* type);
     char* getListedDirectories(char* directory);
@@ -11,4 +12,6 @@
     void* findFile(char* name, char* root, char* type);
     char* ParseFile(FILE* fp, size_t* line);
     WIN32_FIND_DATA* listFiles(char* key, char* path);
+    int dirExists(const char* path);
+    TCHAR* getCurrentDirectory();
 #endif