+++ projects/cfm/cfm.h
@@ -3,6 +3,8 @@
     #include <stdio.h>
     #include <windows.h>
     #include <direct.h>
+    #include <sys/stat.h>
+    #include <stdio.h>
     #include <stringManipulation.h>
 
     FILE* lookInRoot(char* filename, char* type);
@@ -11,4 +13,6 @@
     void* findFile(char* name, char* root, char* type);
     char* ParseFile(FILE* fp, size_t* line);
     WIN32_FIND_DATA* listFiles(char* key, char* path);
+    int dirExists(const char* path);
+    TCHAR* getCurrentDirectory();
 #endif