+++ projects/cl/install.h
@@ -1,0 +1,7 @@
+#ifndef INSTALL_H
+#define INSTALL_H
+    #include <string.h>
+    #include <cfm2.h>
+    #include <stdbool.h>
+    void* clInstall(char* name);
+#endif